Omits Book notes from the homepage list
