+++ src/app/(site)/page.tsx
@@ -1,26 +1,30 @@
-import getNotesAndPosts from "src/data/getNotesAndPosts";
+import { getHomepage, getNotesAndPosts } from "src/data/getHomepage";
 import NoteItem from "src/components/note-item";
 import PostItem from "src/components/post";
 import Reading from "src/components/reading";
 import list from "src/styles/list.module.css";
 
 export default async function HomePage() {
+  const { reading, read } = await getHomepage();
   const notesAndPosts = await getNotesAndPosts();
-
   return (
     <>
-      {/** @ts-expect-error Server Component */}
-      <Reading />
+      <Reading reading={reading} read={read} />
       <ul className={list.noStyle}>
-      {notesAndPosts.map((item) => {
-        const component = item._type === "note" ? <NoteItem note={item} /> : <PostItem post={item} link />;
-        return (
-          <li key={item._id} className={list.listItem}>
-            {component}
-          </li>
-        );
-      })}
-    </ul>
+        {notesAndPosts.map((item) => {
+          const component =
+            item._type === "note" ? (
+              <NoteItem note={item} />
+            ) : (
+              <PostItem post={item} link />
+            );
+          return (
+            <li key={item._id} className={list.listItem}>
+              {component}
+            </li>
+          );
+        })}
+      </ul>
     </>
   );
 }
